+++ CHANGES.md
@@ -8,6 +8,9 @@ Version 0.15.2
 
 To be released.
 
+ -  Fixed a bug of `Object.toJsonLd()` method where it had not fall back to
+    the proper compact form when the heuristic compact form was not available.
+
 
 Version 0.15.1
 --------------
@@ -80,6 +83,15 @@ Released on September 11, 2024.
 
  -  Added `-r`/`--raw` option to `fedify lookup` command to output the raw JSON
     object.
+
+
+Version 0.14.5
+--------------
+
+Released on September 26, 2024.
+
+ -  Fixed a bug of `Object.toJsonLd()` method where it had not fall back to
+    the proper compact form when the heuristic compact form was not available.
 
 
 Version 0.14.4